Extract templates into their own module
