+++ src/templates/index.ts
@@ -1,0 +1,2 @@
+export * from './directoryList.ts'
+export * from './notFound.ts'
